Drop the task which already stopped in restoreTask
In LauncherActivity, find a task to restore, when the task is not
running (id = -1), actually restoring must be failed, the task
we found to restore can not be resued because the topActivity of it
is null, ActivityManager will create a new one, and the task we found
will be leaked.

So just finishAndRemove the stopped task (id = -1), and return
the intent, we will created a new active task to restore it.

Bug: 110440661
Test: make DocumentsUI

Change-Id: Ibe6e77d4d7e7ab0542afffdd98de2540fe72d427
Signed-off-by: mydongistiny <jaysonedson@gmail.com>